vpnkit-forwarder: new host implementation, new protocol
Previously the 9P host side was inside the main `vpnkit` binary
and used a synthetic 9P filesystem. This was sub-optimal because

- we had to implement all the forwarding protocol twice, once
  in OCaml and once in Go
- we had to duplicate many test cases
- 9P is not used anywhere else for control interfaces so it's
  a bit "surprising"
- we needed 2 9P clients: one via a Linux mount and one via a
  Unix domain socket

This patch implements the host side of the port-forwarding system
in Go and replaces the 9P protocol with a simple HTTP based one.

The control plane over HTTP can be exposed over both AF_VSOCK and
as a Unix domain socket/ named pipe for maximum convenience.

The data plane can be connected either host -> VM or VM -> host,
depending on your preferences (and the stability of your AF_VSOCK
implementation)

Signed-off-by: David Scott <dave.scott@docker.com>